feat: currentTimeZoneToIsoString function and implement it
A new utility function named 'currentTimeZoneToIsoString' has been added to 'utils.ts'. This function returns the current time converted to an ISO string. It's implemented in 'form.tsx' to provide a default value for the commit date field and in 'route.ts' to generate the time a commit is created. Also, the handling of an empty commit message has been improved in 'route.ts'.